refactor: proceedRound and print code to Stadium from Game
